+++ source/be_rterror.h
@@ -1,7 +1,9 @@
 #ifndef BE_RTERROR_H_
 #define BE_RTERROR_H_
 
-void RTFatalType(int *pc)
+#include <stdint.h>
+
+void RTFatalType(intptr_t *pc)
 #ifdef EUNIX
 __attribute__ ((noreturn))
 #endif
@@ -32,7 +34,7 @@ object_ptr BiggerStack();
 void atom_condition();
 void MainScreen();
 void RangeReading(object subs, int len);
-void BadSubscript(object subs, long length);
+void BadSubscript(object subs, int length);
 void NoValue(symtab_ptr s);
 
 #endif
